configured routes and created the Homapage
